Split Auth functionality out to new file
